upgrade deps, better docs, script tweeks
The upgrade of the dependencies was just a run of `dependencies-upgrade`.

The `org2gfm` script did change in a few ways:

    - dropped the `--no-nix` switch which wasn't useful
    - put in `--yes` and `--no` option for evaluation queries
    - defaulted `:eval` to `no-export`

Also, I made the derivation of the scripts have meta.description attributes, so
that they work well with `nix search`.

Most of these changes are improved docs.  Changing the defaulting of `:eval`
simplified how Org-mode is used.  And this is documented now in
`doc/org2gfm-design.org`.  Also, there's some documentation of Nix
(`doc/nix.org`) that I can copy to other projects.