ORG2GFM: Remove Markdown file before evaluation
Sometimes Emacs fails to evaluate a file, but silently exits.  This can
make it seem like the file was successfully regenerated, but we're still
looking at the old version of the file.  This is confusing, and hard to
detect.

Since these files should always be under source control, and because the
evaluation is going to destroy the old version ultimately anyway, it
seems reasonable to delete the Markdown file just before evaluation.